removed broadcasted stuff. leave and end event now work. Ricardo fixed the event creation error.
