+++ app/src/main/res/layout/activity_settings.xml
@@ -39,13 +39,5 @@
         android:layout_marginTop="23dp"
         android:onClick="toAbout"/>
 
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/NoRole_log_out_button"
-        android:id="@+id/log_out_button"
-        android:layout_below="@+id/about_button"
-        android:layout_centerHorizontal="true"
-        android:onClick="toLogOut" />
 
 </RelativeLayout>
